0.3.1: Fix two bugs, add optional txids=[] for existing txids, improve testing
So if we were getting unsorted unspents before, we were trying to
be smart on speed and break when the confirmations got out of hand.
Now we just continue in case it's unsorted.
(No test case for this yet)

I ran into a bug where payments with identical amounts that were
outside of that 6 confirmation window were being skipped.

I had my existing TXID logic outside of this, so it would return
a single TXID that would match the one already paid. The second
payment would get ignored. This is fixed if you move your existing
TXID logic into bitcoinacceptor. Tests added for this.